requirements.txt
drf_yasg==1.17.1
Django==2.2.4
djangorestframework_simplejwt==4.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
whitenoise==5.1.0
dj-database-url==0.5.0
psycopg2==2.8.5
django-filter==2.3.0